refactor(discord): Move utilities to seperate files
