+++ src/discord/util/formats.ts
@@ -1,0 +1,4 @@
+export function toDiscordTimestamp(date: Date): string {
+    const epoch = (date.getTime() / 1000).toFixed(0)
+    return `<t:${epoch}>`
+}
