+++ src/discord/util/index.ts
@@ -1,0 +1,5 @@
+import { Theme, User } from "@prisma/client"
+import { BaseCommandInteraction, ButtonInteraction } from "discord.js"
+
+export type DatabaseUserWithThemes = User & { themes: Theme[] }
+export type MessageBasedInteraction = ButtonInteraction | BaseCommandInteraction
